backend/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
applicationinsights==0.11.10
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
azure-core==1.35.0
azure-communication-email==1.0.0
azure-identity==1.25.0
azure-monitor-opentelemetry-exporter==1.0.0b41
bcrypt==4.1.2
billiard==4.2.4
cachetools==5.5.2
celery==5.3.4
-e git+https://github.com/RakeshMahendran/Saramsa.git@a20af6b2a6cb0143e04ff4f5acb0f300317a3bce#egg=celery_ops&subdirectory=celery_ops
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.7
dateparser==1.2.1
Deprecated==1.2.18
distro==1.9.0
Django==4.2.20
django-cors-headers==4.7.0
dj-database-url==2.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.27.1
exceptiongroup==1.3.0
fastapi==0.128.0
filelock==3.20.0
fixedint==0.1.6
flower==2.0.1
fsspec==2025.12.0
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
h11==0.14.0
hf-xet==1.2.0
httpcore==1.0.8
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.3.5
humanize==4.15.0
hypothesis==6.151.4
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
isodate==0.7.2
Jinja2==3.1.3
jiter==0.9.0
joblib==1.5.3
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.6.2
MarkupSafe==3.0.2
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
msrest==0.7.1
networkx==3.6.1
numpy==2.4.1
oauthlib==3.3.1
openai==1.75.0
opentelemetry-api==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-instrumentation-dbapi==0.58b0
opentelemetry-instrumentation-django==0.58b0
opentelemetry-instrumentation-psycopg2==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-wsgi==0.58b0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
packaging==25.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==4.25.5
psutil==7.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
starlette==0.50.0
sympy==1.13.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.22.2
torch==2.6.0+cu124
tornado==6.5.4
tqdm==4.67.1
transformers==5.0.0
typer-slim==0.21.1
typing-inspection==0.4.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.40.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==16.0
wrapt==1.17.2
zipp==3.23.0
